move common funcs to helpers
