+++ analysis/statistics/_helpers.py
@@ -2,6 +2,10 @@ import os
 from typing import List, Optional, Union, Any
 
 import pandas as pd
+import scipy.stats as stats
+import scikit_posthocs as sp
+import plotly.graph_objects as go
+from plotly.subplots import make_subplots
 
 import pEYES as peyes
 from pEYES._DataModels.UnparsedEventLabel import UnparsedEventLabelType, UnparsedEventLabelSequenceType
@@ -45,3 +49,144 @@ def get_data_impl(
     if sub_index:
         data = data.loc[sub_index]
     return data
+
+
+def statistical_analysis(
+        data: pd.DataFrame,
+        gt_cols: List[str],
+        multi_comp: Optional[str] = "fdr_bh",
+):
+    """
+    For each unique value in the input DataFrame's index and each of the GT labelers, performs Kruskal-Wallis test with
+    post-hoc Dunn's test for multiple comparisons. Returns the KW-statistic, KW-p-value, Dunn's-p-values and number of
+    samples for each (index, GT labeler) pair.
+
+    :param data: DataFrame; Should have the following MultiIndex structure:
+        - Index :: single level
+        - Columns :: 1st level: trial id
+        - Columns :: 2nd level: GT labeler
+        - Columns :: 3rd level: Pred labeler (detection algorithm)
+    :param gt_cols: List of GT labelers to compare.
+    :param multi_comp: Method for multiple comparisons correction when performing Dunn's post-hoc test.
+
+    :return: Four DataFrames containing results for the statistical analysis.
+        - statistics: KW statistic; index is metric name, columns are GT labelers.
+        - pvalues: KW p-value; index is metric name, columns are GT labelers.
+        - dunns: DataFrame with Dunn's p-values for pair-wise comparisons; index and columns are Pred labelers.
+        - Ns: Number of data-points (trials) for each (metric, GT labeler, Pred labeler) pair; index is metric name,
+            columns multiindex with pairs of (GT, Pred) labelers.
+    """
+    indices = sorted(
+        data.index.unique(),
+        key=lambda met: u.METRICS_CONFIG[met][1] if met in u.METRICS_CONFIG else ord(met[0])
+    )
+    statistics, pvalues, dunns, Ns = {}, {}, {}, {}
+    for i, idx in enumerate(indices):
+        for j, gt_col in enumerate(gt_cols):
+            gt_series = data.xs(gt_col, level=u.GT_STR, axis=1).loc[idx]
+            gt_df = gt_series.unstack().drop(columns=gt_cols, errors='ignore')
+            detector_values = {col: gt_df[col].explode().dropna().values for col in gt_df.columns}
+            statistic, pvalue = stats.kruskal(*detector_values.values(), nan_policy='omit')
+            dunn = pd.DataFrame(
+                sp.posthoc_dunn(a=list(detector_values.values()), p_adjust=multi_comp).values,
+                index=gt_df.columns, columns=gt_df.columns
+            )
+            statistics[(idx, gt_col)] = statistic
+            pvalues[(idx, gt_col)] = pvalue
+            dunns[(idx, gt_col)] = dunn
+            Ns.update({(idx, gt_col, det): det_vals.shape[0] for det, det_vals in detector_values.items()})
+
+    # create outputs
+    statistics = pd.Series(statistics).unstack()
+    pvalues = pd.Series(pvalues).unstack()
+    statistics.index.names = pvalues.index.names = [peyes.constants.METRIC_STR]
+    statistics.columns.names = pvalues.columns.names = [u.GT_STR]
+    dunns = pd.Series(dunns).unstack()
+    dunns.index.names = dunns.columns.names = [u.PRED_STR]
+    Ns = pd.Series(Ns)
+    Ns.index.names = [peyes.constants.METRIC_STR, u.GT_STR, u.PRED_STR]
+    Ns = Ns.unstack([u.GT_STR, u.PRED_STR])
+    return statistics, pvalues, dunns, Ns
+
+
+def distributions_figure(
+        data: pd.DataFrame,
+        gt1: str,
+        title: str,
+        gt2: Optional[str] = None,
+        only_box: bool = False,
+) -> go.Figure:
+    """
+    Creates a violin/box subplot for each unique value in the input DataFrame's index. Each subplot (index value)
+    contains violins/boxes for all predictor labelers (detection algorithms), compared to the specified GT labeler(s).
+    If two GT labelers are provided, the plot will show GT1 in the positive (right) side and GT2 in the negative (left) side.
+
+    :param data: DataFrame; Should have the following MultiIndex structure:
+        - Index :: single level
+        - Columns :: 1st level: trial id
+        - Columns :: 2nd level: GT labeler
+        - Columns :: 3rd level: detector
+    :param gt1: name of the first GT labeler to compare.
+    :param title: optional; title for the plot.
+    :param gt2: optional; name of the second GT labeler to compare.
+    :param only_box: if True, only box plots will be shown.
+
+    :return: Plotly figure with the violin/box plot.
+    """
+    gt_cols = list(filter(None, [gt1, gt2]))
+    assert 0 < len(gt_cols) <= 2
+    indices = sorted(
+        data.index.unique(),
+        key=lambda met: u.METRICS_CONFIG[met][1] if met in u.METRICS_CONFIG else ord(met[0])
+    )
+    ncols = 1 if len(indices) <= 3 else 2
+    nrows = len(indices) if len(indices) <= 3 else sum(divmod(len(indices), ncols))
+    fig = make_subplots(
+        rows=nrows, cols=ncols,
+        shared_xaxes=False,
+        subplot_titles=list(map(lambda met: u.METRICS_CONFIG[met][0] if met in u.METRICS_CONFIG else met, indices)),
+    )
+    for i, idx in enumerate(indices):
+        r, c = (i, 0) if ncols == 1 else divmod(i, ncols)
+        for j, gt_col in enumerate(gt_cols):
+            gt_series = data.xs(gt_col, level=u.GT_STR, axis=1).loc[idx]
+            gt_df = gt_series.unstack().drop(columns=gt_cols, errors='ignore')
+            for d, detector in enumerate(gt_df.columns):
+                det_name = detector.removesuffix("Detector")
+                if len(gt_cols) == 1:
+                    violin_side = None
+                    opacity = 0.75
+                else:
+                    violin_side = 'positive' if j == 0 else 'negative'
+                    opacity = 0.75 if j == 0 else 0.25
+                if only_box:
+                    fig.add_trace(
+                        row=r + 1, col=c + 1,
+                        trace=go.Box(
+                            x0=det_name, y=gt_df[detector].explode().dropna().values,
+                            name=f"{gt_col}, {det_name}", legendgroup=det_name,
+                            marker_color=u.DEFAULT_DISCRETE_COLORMAP[d], line_color=u.DEFAULT_DISCRETE_COLORMAP[d],
+                            opacity=opacity, boxmean='sd', showlegend=i == 0,
+                        )
+                    )
+                else:
+                    fig.add_trace(
+                        row=r + 1, col=c + 1,
+                        trace=go.Violin(
+                            x0=det_name, y=gt_df[detector].explode().dropna().values,
+                            side=violin_side, opacity=opacity, spanmode='hard',
+                            fillcolor=u.DEFAULT_DISCRETE_COLORMAP[d],
+                            name=f"{gt_col}, {det_name}", legendgroup=det_name, scalegroup=idx, showlegend=i == 0,
+                            box_visible=True, meanline_visible=True, line_color='black',
+                        ),
+                    )
+        y_range = u.METRICS_CONFIG[idx][2] if idx in u.METRICS_CONFIG else None
+        fig.update_yaxes(row=r + 1, col=c + 1, range=y_range)
+    fig.update_layout(
+        title=title,
+        violinmode='overlay',
+        boxmode='group',
+        boxgroupgap=0,
+        boxgap=0,
+    )
+    return fig
